Move common functions to common file
